requirements.txt
spidev
RPi.GPIO
websockets
